// src/Types/Enums.ts
export enum TIME_FORMAT {
	TWELVE_HOUR = "12-hour",
	TWENTY_FOUR_HOUR = "24-hour",
}

/** Locale of the date shown under the clock. */
export enum DATE_LANGUAGE {
	FRENCH = "fr-FR",
	ENGLISH = "en-GB",
}

export enum BOOKMARK_SOURCE {
	ALL = "all",
	GROUP = "group",
}

export enum QUOTE_SOURCE {
	QUOTEABLE = "Quoteable",
	MY_QUOTES = "My quotes",
	BOTH = "Both",
}